flagging unmaintained datetime pickers (#1772)
I keep going back to this list for datepicker libraries. Two first options on the list are not maintained for almost a year and over a year respectively. I think it is useful to flag them out so that you won't go to those repositories only to find out that the packages are not maintained anymore. 

A good rule of thumb for flagging unmaintained packages was proposed [here](https://github.com/vuejs/awesome-vue/issues/1489) - we can consider a package unmaintained if it hasn't been updated in the last 6 months.